+++ src/pages/api/auth/login.ts
@@ -1,6 +1,6 @@
 export const prerender = false;
 import type { APIRoute } from "astro";
-import { supabase } from "../../../../utils/supabase";
+import { supabase } from "../../../lib/supabase";
 
 export const POST: APIRoute = async ({ request, cookies, redirect }) => {
 	const formData = await request.formData();
