+++ src/pages/index.astro
@@ -2,6 +2,41 @@ export const prerender = false
 ---
 import "../assets/app.css"
 import Login from "../components/login.astro"
+import { supabase } from "../lib/supabase";
+
+const accessToken = Astro.cookies.get("sb-access-token");
+const refreshToken = Astro.cookies.get("sb-refresh-token");
+
+if (!accessToken || !refreshToken) {
+  return Astro.redirect("/login");
+}
+
+let session;
+try {
+  session = await supabase.auth.setSession({
+    refresh_token: refreshToken.value,
+    access_token: accessToken.value,
+  });
+  if (session.error) {
+    Astro.cookies.delete("sb-access-token", {
+      path: "/",
+    });
+    Astro.cookies.delete("sb-refresh-token", {
+      path: "/",
+    });
+    return Astro.redirect("/login");
+  }
+} catch (error) {
+  Astro.cookies.delete("sb-access-token", {
+    path: "/",
+  });
+  Astro.cookies.delete("sb-refresh-token", {
+    path: "/",
+  });
+  return Astro.redirect("/login");
+}
+
+const email = session.data.user?.email;
 ---
 
 <html lang="en">
@@ -14,7 +49,6 @@ import Login from "../components/login.astro"
 	</head>
 	<body "h-screen">
 	  <div class="flex justify-center items-center h-screen">
-	  <Login/>
 	  </div>
 	</body>
 </html>
